Let there be config cli
There are many grpc types we redefine and provide From impls for, this
is mostly for Hex fields that come in as integers, and String fields
that come in as Byte Vecs.

We write these to files.
To update, users will change the file and push. It helps to have those
values in understandable formats, and we can use this as an opportunity
to do extra validation.